Python_Flask_FastApi/requirements.txt
altair
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
filelock
Flask
flask-cors
fonttools
fsspec
gitdb
GitPython
huggingface-hub
idna
imageio
itsdangerous
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
lazy_loader
lxml
MarkupSafe
matplotlib
mpmath
narwhals
networkx
ninja
numpy
opencv-python
opencv-python-headless
packaging
pandas
pdfminer.six
pdfplumber
pillow
protobuf
pyarrow
pyclipper
pycparser
pydeck
pydub
pyparsing
PyPDF2
pypdfium2
pytesseract
python-bidi
python-dateutil
python-docx
pytz
PyYAML
pyzbar
referencing
regex
requests
rpds-py
safetensors
scikit-image
scikit-learn
scipy
setuptools
shapely
six
smmap
streamlit
sympy
tenacity
threadpoolctl
tifffile
tokenizers
toml
torch
torchvision
tornado
tqdm
transformers
typing_extensions
tzdata
urllib3
watchdog
Werkzeug
blinker
click
colorama
Flask
flask-cors
itsdangerous
Jinja2
MarkupSafe
piexif
pillow
Werkzeug
